Extract type and declaration modules for statements
